audio2lyrics/requirements.txt
fastapi
uvicorn[standard]
numpy
ffmpeg-python
openai-whisper
torch
python-multipart